+++ src/tests/mod.rs
@@ -8,6 +8,7 @@
 )]
 
 mod arg_serializer;
+mod array_commands;
 mod bitmap_commands;
 mod bloom_commands;
 mod buffer_decoder;